+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 - Add serverless build lambda function
 - Define project structure
 - Version 1: lambda trigger event from test
+
+## [0.0.2] - 2025-07-17
+
+### Added
+- Update s3 event trigger lambda
+- Version 2: lambda trigger event from s3 notification